creat nft recipient account fix, startup script update (#102)

+++ run.sh
@@ -1,21 +1,10 @@
-#!/usr/bin/env bash
-
-echo "Clearing previous files..."
-rm -rf ~/.mp*
-mkdir -p ~/.mpd/config
-cp config.toml ~/.mpd/config
-
-echo "Building..."
-make install
-
-echo "Initialization..."
-
-cur_path=$( cd "$(dirname "${BASH_SOURCE[0]}")" ; pwd -P )
+#!/usr/bin/env zsh
 
 account_num=200
 money_count="100000token"
-file_output=$cur_path"/out.txt"
+file_output="out.txt"
 stake_count=100000000
+PSW="12345678"
 
 while test $# -gt 0; do
   case "$1" in
@@ -26,12 +15,14 @@ while test $# -gt 0; do
       echo " "
       echo "options:"
       echo "-h, --help                show brief help"
-      echo "--demo                    set demo mode (add demo accounts)"
+      echo "--build                   manual rebuild marketplace (not for docker)"
+      echo "--reset                   force reset blockchain state (clear all data)"
+      echo "--demo                    add demo accounts (used only with --reset)"
       echo "-n, --num_account=n       specify number of demo accounts | 200 default"
       echo "-m, --money=m             specify token amount for demo account | 100000token default"
       echo "-s, --stake=s             specify stake amount for demo account | 100000000 default"
       echo "-o, --output_file=o       specify output file | out.txt default"
-      echo "--embeded                 set embeded mode (for docker)"
+      echo "--norun                   omit service start"
       exit 0
       ;;
     -n|--num_account)
@@ -74,12 +65,20 @@ while test $# -gt 0; do
       fi
       shift
       ;;
-    --embeded)
-      EMBEDED=true
+    --build)
+      BUILD=true
       shift
       ;;
     --demo)
       DEMO=true
+      shift
+      ;;
+    --reset)
+      RESET=true
+      shift
+      ;;
+    --norun)
+      NORUN=true
       shift
       ;;
     *)
@@ -88,54 +87,61 @@ while test $# -gt 0; do
   esac
 done
 
-rm $file_output
-
-mpd init node0 --chain-id mpchain
-
-echo "Adding keys..."
-
-mpcli keys add user1 <<< "12345678"
-mpcli keys add user2 <<< "12345678"
-mpcli keys add sellerBeneficiary <<< "12345678"
-mpcli keys add buyerBeneficiary <<< "12345678"
-mpcli keys add dgaming <<< "12345678"
-
-mpd add-genesis-account $(mpcli keys show user1 -a) 1000token,100000000stake
-mpd add-genesis-account $(mpcli keys show user2 -a) 1000token,100000000stake
-mpd add-genesis-account $(mpcli keys show sellerBeneficiary -a) 1000token,100000000stake
-mpd add-genesis-account $(mpcli keys show buyerBeneficiary -a) 1000token,100000000stake
-mpd add-genesis-account $(mpcli keys show dgaming -a) 1000token,100000000stake
-
-
-if [[ $DEMO ]]; then
-  for ((i=1;i<=$account_num;i++));
-  do
-    pwd=$(gpg --gen-random --armor 1 14)
-    mnemonic=$(mpcli keys add demo$i <<< $pwd 2>&1 >/dev/null | tail -1)
-
-    mpd add-genesis-account $(mpcli keys show demo$i -a) $money_count,${stake_count}stake
-    echo "demo$i      $pwd        $money_count   ${stake_count}stake       $mnemonic" >> $file_output
-  done
+if [[ $BUILD ]]; then
+  echo "Building..."
+  make install
 fi
 
+if [[ $RESET ]]; then
+  echo "Clearing previous files..."
+  rm -rf ~/.mp*
+fi
 
-echo "Adding genesis account..."
+if [ ! -f ~/.mpd/config/config.toml ]; then
+  mkdir -p ~/.mpd/config
+  cp config.toml ~/.mpd/config
 
-echo "Configuring..."
-mpcli config chain-id mpchain
-mpcli config output json
-mpcli config indent true
-mpcli config trust-node true
+  echo "Initialization..."
+  mpd init node0 --chain-id mpchain
 
-mpd gentx --name user1 <<< "12345678"
-mpd collect-gentxs
-mpd validate-genesis
+  echo "Adding genesis accounts..."
+  mpcli keys add user1 -i <<< $PSW < data_u1.txt
+  mpcli keys add user2 -i <<< $PSW < data_u2.txt
+  mpcli keys add sellerBeneficiary -i <<< $PSW < data_sb.txt
+  mpcli keys add buyerBeneficiary -i <<< $PSW < data_bb.txt
+  mpcli keys add dgaming -i <<< $PSW < data_dg.txt
 
-if [[ $EMBEDED ]]; then
-  sed -i 's/proxy_app = "tcp:\/\/127.0.0.1:26658"/proxy_app = "tcp:\/\/0.0.0.0:26658"/' /root/.mpd/config/config.toml
-  sed -i 's/laddr = "tcp:\/\/127.0.0.1:26657"/laddr = "tcp:\/\/0.0.0.0:26657"/' /root/.mpd/config/config.toml
-else
+  mpd add-genesis-account $(mpcli keys show user1 -a) 999999token,100000000stake
+  mpd add-genesis-account $(mpcli keys show user2 -a) 999999token,100000000stake
+  mpd add-genesis-account $(mpcli keys show sellerBeneficiary -a) 1000token,100000000stake
+  mpd add-genesis-account $(mpcli keys show buyerBeneficiary -a) 1000token,100000000stake
+  mpd add-genesis-account $(mpcli keys show dgaming -a) 1000token,100000000stake
+
+  if [[ $DEMO ]]; then
+    rm $file_output
+    for ((i=1;i<=$account_num;i++));
+    do
+      pwd=$(gpg --gen-random --armor 1 14)
+      mnemonic=$(mpcli keys add demo$i <<< $pwd 2>&1 >/dev/null | tail -1)
+
+      mpd add-genesis-account $(mpcli keys show demo$i -a) $money_count,${stake_count}stake
+      echo "demo$i      $pwd        $money_count   ${stake_count}stake       $mnemonic" >> $file_output
+    done
+  fi
+
+  echo "Configuring..."
+  mpcli config chain-id mpchain
+  mpcli config output json
+  mpcli config indent true
+  mpcli config trust-node true
+
+  mpd gentx --name user1 <<< $PSW
+  mpd collect-gentxs
+  mpd validate-genesis
+fi
+
+if [[ -z $NORUN ]]; then
   echo "Starting node..."
-
+  mpcli rest-server --chain-id mpchain --trust-node --laddr tcp://0.0.0.0:1317 > /dev/null &
   mpd start
 fi
